+++ moviepy/editor.py
@@ -6,7 +6,7 @@ MoviePy by simply typing:
 
 In particular it will load many effects from the video.fx and audio.fx
 folders and turn them into VideoClip methods, so that instead of
->>> clip.fx( vfx.resize, 2 ) # or equivalently vfx.resize(clip, 2)
+>>> clip.fx( vfx.resize, 2 ) or equivalently vfx.resize(clip, 2)
 we can write
 >>> clip.resize(2)
 
@@ -14,11 +14,36 @@ It also starts a PyGame session (if PyGame is installed) and enables
 clip.preview().
 """
 
+__all__ = [
+    "afx",
+    "AudioClip",
+    "AudioFileClip",
+    "BitmapClip",
+    "clips_array",
+    "ColorClip",
+    "CompositeAudioClip",
+    "CompositeVideoClip",
+    "concatenate_audioclips",
+    "concatenate_videoclips",
+    "convert_to_seconds",
+    "download_webfile",
+    "ffmpeg_tools",
+    "ImageClip",
+    "ImageSequenceClip",
+    "ipython_display",
+    "TextClip",
+    "transfx",
+    "vfx",
+    "VideoClip",
+    "VideoFileClip",
+    "videotools",
+]
+
 # Note that these imports could have been performed in the __init__.py
 # file, but this would make the loading of moviepy slower.
 
 import os
-import sys
+import inspect
 
 
 # Hide the welcome message from pygame: https://github.com/pygame/pygame/issues/542
@@ -37,8 +62,8 @@ from .audio.io.AudioFileClip import AudioFileClip
 
 # FX
 
-import moviepy.video.fx.all as vfx
-import moviepy.audio.fx.all as afx
+import moviepy.video.fx as vfx
+import moviepy.audio.fx as afx
 import moviepy.video.compositing.transitions as transfx
 
 # Tools
@@ -50,44 +75,26 @@ from .tools import convert_to_seconds
 
 try:
     from .video.io.sliders import sliders
+
+    __all__.append("sliders")
 except ImportError:
     pass
 
-# The next loop transforms many effects into VideoClip methods so that
-# they can be called with myclip.resize(width=500) instead of
-# myclip.fx( vfx.resize, width= 500)
-for method in [
-    "afx.audio_fadein",
-    "afx.audio_fadeout",
-    "afx.audio_normalize",
-    "afx.volumex",
-    "transfx.crossfadein",
-    "transfx.crossfadeout",
-    "vfx.crop",
-    "vfx.fadein",
-    "vfx.fadeout",
-    "vfx.invert_colors",
-    "vfx.loop",
-    "vfx.margin",
-    "vfx.mask_and",
-    "vfx.mask_or",
-    "vfx.resize",
-    "vfx.rotate",
-    "vfx.speedx",
-]:
+# Transforms the effects into Clip methods so that
+# they can be called with clip.resize(width=500) instead of
+# clip.fx(vfx.resize, width=500)
+audio_fxs = inspect.getmembers(afx, inspect.isfunction)
+video_fxs = (
+    inspect.getmembers(vfx, inspect.isfunction)
+    + inspect.getmembers(transfx, inspect.isfunction)
+    + audio_fxs
+)
 
-    exec("VideoClip.%s = %s" % (method.split(".")[1], method))
+for name, function in video_fxs:
+    setattr(VideoClip, name, function)
 
-
-for method in [
-    "afx.audio_fadein",
-    "afx.audio_fadeout",
-    "afx.audio_loop",
-    "afx.audio_normalize",
-    "afx.volumex",
-]:
-
-    exec("AudioClip.%s = %s" % (method.split(".")[1], method))
+for name, function in audio_fxs:
+    setattr(AudioClip, name, function)
 
 
 # adds easy ipython integration