+++ moviepy/video/compositing/transitions.py
@@ -11,6 +11,9 @@ from moviepy.video.fx.fadeout import fadeout
 from .CompositeVideoClip import CompositeVideoClip
 
 
+__all__ = ["crossfadein", "crossfadeout", "slide_in", "slide_out", "make_loopable"]
+
+
 @requires_duration
 @add_mask_if_none
 def crossfadein(clip, duration):